TyperAssistant.from_id deprecated, so workaround
I have discovered a bug with TyperAssistant.from_id that causes it to
fail due to the way the base class implements 'name'. There's a few ways
this can be addressed but I believe the cleanest way is to rethink the
use of polymorphism in its entirety for this use case. This means a true
fix will likely need to wait for a refactor that removes the Assistant
base class from the codebase more or less in its entirety and replaces
it with functions that act on OpenAI resources and wrap OpenAI calls,
with the TyperAssistant class being the base class which wraps an app
and an assistant together.

This refactor can't fit in 0.3 so for now I'm adding
TyperAssistant.from_id_with_app which is a workaround that allows
essentially the same functionality as Assistant.from_id but tied to a
typer.Typer app. This is an early warning that 0.4 is likely to once
again significantly refactor the API. Sorry for any problems this
creates!